+++ app/helpers/sort_helper.rb
@@ -7,7 +7,7 @@ module SortHelper
     if params[:sort_dir] && params[:sort]
       sort_people_skills
     else
-      @search_results
+      default_sort_people_skills
     end
   end
 
@@ -58,6 +58,15 @@ module SortHelper
                         sort_by_people_skill_attr
                       end
     params[:sort_dir] == 'asc' ? @search_results : @search_results.reverse
+  end
+
+  def default_sort_people_skills
+    @search_results.sort_by do |person_skills|
+      [
+        -person_skills.length,
+        person_skills.map { |people_skill| people_skill.skill.title.downcase }.sort
+      ]
+    end
   end
 
   def sort_by_human_attr